+++ src/main/java/com/els/bean/RoomInfo.java
@@ -6,7 +6,7 @@ public class RoomInfo {
 
 	private Integer roomid;
 	// 0未开始 1：游戏中 2:游戏结束
-	private byte roomStatus;
+	private Integer roomStatus;
 	// 0房主 1：玩家 2:旁观者
 	private Integer userStatus;
 	private List<JhddUsers> userList;
@@ -19,11 +19,11 @@ public class RoomInfo {
 		this.roomid = roomid;
 	}
 
-	public byte getRoomStatus() {
+	public Integer getRoomStatus() {
 		return roomStatus;
 	}
 
-	public void setRoomStatus(Byte roomstate) {
+	public void setRoomStatus(Integer roomstate) {
 		this.roomStatus = roomstate;
 	}
